Restructure and new design
Moved dialog timer to separate folder. Moved show dialog (onDone) to separate function. Design changes
